requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
redis
qdrant-client